refactor: utils is now a folder
