+++ betterrolls5e/scripts/betterrolls5e.js
@@ -1,6 +1,6 @@
 import { DND5E } from "../../../systems/dnd5e/module/config.js";
 import { CustomRoll, CustomItemRoll } from "./custom-roll.js";
-import { i18n, Utils, ItemUtils } from "./utils.js";
+import { i18n, Utils, ItemUtils } from "./utils/index.js";
 import { getSettings } from "./settings.js";
 
 // Returns whether an item makes an attack roll
