+++ betterrolls5e/scripts/fields.js
@@ -1,4 +1,4 @@
-import { dnd5e, i18n, ActorUtils, ItemUtils, Utils } from "./utils.js";
+import { dnd5e, i18n, ActorUtils, ItemUtils, Utils } from "./utils/index.js";
 import { BRSettings, getSettings } from "./settings.js";
 
 /**
