+++ betterrolls5e/scripts/utils/index.js
@@ -1,0 +1,3 @@
+export * from "./dice-collection.js";
+export * from "./proxy.js";
+export * from "./utils.js";
